Restore values after refreshing the selects
Up until now because the selects options are completely regenerated
every time we change the builder options or changed the selected hero
the user always ended with None selected for all of them.

Just store the current value in a variable and apply it again after the
select is populated.
We also skip checking legality or max scoring even when enabled since
the user might have enabled those options just to select a skill and
disabled them again for convenience when going through the rest