Initial ascendent boons support
Essentially works as an additional boon that doesn't affect the order of
boosts given by dragonflowers or merges.

The only catch is on first neutral merge where an ascendent boon cannot
receive the extra stat and thus is passed to the fourth higher stat.

The clean icon for the floret is not available yet so that will come
later (should render besides the rarity line)